requirements.txt
flask==1.0.2
passlib==1.7.1
pyjwt==1.7.1